ch1: thru first tests passing
Curiousity about go golang was piqued by how functions are incorporated
into structs. ChatGPT called this something like "recieving." I like the
TDD aspect of the book.

It also took me a while to initialize this as a go module... I didn't
realize the first segment had to be a domain.