Begin creating the AST
Define different node types, such as program
(root) and statements (beginning with the `let`
statement).

Golang question - is there such a thing as a
"class constant?" For the `LetStatement` struct,
can the token be a value rather than a property?

It also looks like there are mechanisms for inheritance in
go. It would be good to get a basic handling of those things.